Refactor Receiver
