--- src/contactmanagementsoftware/command/CancelReceiver.java
@@ -1,9 +1,0 @@
-package contactmanagementsoftware.command;
-
-public class CancelReceiver extends Receiver {
-
-    public void  cancel() {
-        jPanel1.setVisible(true);
-        jPanel3.setVisible(false);
-    }
-}
