+++ src/contactmanagementsoftware/command/ContactReceiver.java
@@ -1,16 +1,35 @@
 package contactmanagementsoftware.command;
 
 import contactmanagementsoftware.*;
+import org.jdesktop.swingx.JXTable;
 
 import javax.swing.*;
+import java.io.File;
+import java.io.IOException;
 import java.util.ArrayList;
 
-public class AddContactReceiver extends Receiver {
+public class ContactReceiver {
+    static MUI mui;
+    JList jList1;
+    JXTable jXTable1;
+    JPanel jPanel1, jPanel2, jPanel3;
+    ArrayList<ArrayList<Acquaintances>> a, temp;
+    boolean flag;
+    int x;
+
+    public ContactReceiver(){
+        mui = mui.getInstance();
+        jList1 = mui.getJList1();
+        jPanel1 = mui.getPanel1();
+        jPanel2 = mui.getPanel2();
+        jPanel3 = mui.getPanel3();
+        jXTable1 = mui.getjXTable1();
+        a = mui.getA();
+    }
 
     public void addContact(){
-        int x = mui.getX();
-        boolean flag = mui.getFlag();
-        ArrayList<ArrayList<Acquaintances>> a = mui.getA();
+        x = mui.getX();
+        flag = mui.getFlag();
         JTextField name = mui.getContactName();
         JTextField mobile = mui.getContactMobile();
         JTextField email = mui.getContactEmail();
@@ -150,10 +169,166 @@ public class AddContactReceiver extends Receiver {
                 break;
             default:
                 break;
-        
+
         }
         jPanel1.setVisible(true);
         jPanel3.setVisible(false);
         mui.setUpTableData();
     }
+
+    public void deleteContact(){
+        int index = jList1.getSelectedIndex();
+        if(index<0){
+            JOptionPane.showMessageDialog(mui, "Select a category!");
+            return;
+        }
+
+        int tindex = jXTable1.getSelectedRow();
+        if(tindex < 0){
+            JOptionPane.showMessageDialog(mui, "Select an entry!");
+            return;
+        }
+
+        int n = JOptionPane.showConfirmDialog(
+                mui,
+                "Are you sure you want to delete this?",
+                "Confirm",
+                JOptionPane.YES_NO_OPTION);
+        if(n==0){
+            a.get(index).remove(tindex);
+            JOptionPane.showMessageDialog(mui, "Successfully Deleted");
+            mui.setUpTableData();
+        }
+    }
+
+    public void editContact(){
+        int index = jList1.getSelectedIndex();
+        if(index<0){
+            JOptionPane.showMessageDialog(mui, "Select a category!");
+            return;
+        }
+        int tindex = jXTable1.getSelectedRow();
+        if(tindex < 0){
+            JOptionPane.showMessageDialog(mui, "Select an entry!");
+            return;
+        }
+
+        mui.setNum(tindex);
+        mui.setFlag(false);
+        mui.setDFlag(false);
+        mui.setX(index);
+        mui.setDescription();
+        jPanel1.setVisible(false);
+        jPanel3.setVisible(true);
+    }
+
+    public void viewFullDetail(){
+
+        int index = jList1.getSelectedIndex();
+        if(index<0){
+            JOptionPane.showMessageDialog(mui, "Select a category!");
+            return;
+        }
+        int tindex = jXTable1.getSelectedRow();
+        if(tindex < 0){
+            JOptionPane.showMessageDialog(mui, "Select an entry!");
+            return;
+        }
+
+        mui.setNum(tindex);
+        mui.setFlag(false);
+        mui.setX(index);
+        jPanel1.setVisible(false);
+        jPanel3.setVisible(true);
+        mui.setDFlag(true);
+        mui.setDescription();
+
+    }
+
+    public void searchContact() {
+        JTextPane details = mui.getDetails();
+
+        String s = (String) JOptionPane.showInputDialog(
+                mui,
+                "Enter name: ",
+                "Input",
+                JOptionPane.PLAIN_MESSAGE,
+                null,
+                null,
+                "");
+        if(s==null)
+            return;
+        jPanel1.setVisible(false);
+        jPanel2.setVisible(true);
+        mui.setStr(s);
+        details.setContentType( "text/html" );
+        mui.runn();
+    }
+
+    public void uploadContacts(){
+        JFileChooser fileChooser = new JFileChooser();
+        fileChooser.setCurrentDirectory(new File(System.getProperty("user.dir")));
+        int result = fileChooser.showOpenDialog(mui);
+        if (result == JFileChooser.APPROVE_OPTION) {
+            File selectedFile = fileChooser.getSelectedFile();
+            try {
+                mui.setTemp((ArrayList<ArrayList<Acquaintances>>) SerializationUtil.deserialize(selectedFile));
+                temp = mui.getTemp();
+            }
+            catch (ClassNotFoundException | IOException e) {
+                JOptionPane.showMessageDialog(mui, "Error");
+                return;
+            }
+        }
+        else{
+            return;
+        }
+        try{
+            for(int i = 0; i < 4; i++){
+                for(int j = 0; j < temp.get(i).size(); j++){
+                    a.get(i).add(temp.get(i).get(j));
+                }
+            }
+        }
+        catch(Exception e){
+
+        }
+        mui.setUpTableData();
+    }
+
+    public void saveContacts(){
+        String s = (String) JOptionPane.showInputDialog(
+                mui,
+                "Enter file name: (*.ser)",
+                "Input",
+                JOptionPane.PLAIN_MESSAGE,
+                null,
+                null,
+                "output.ser");
+        if(s==null)
+            return;
+        if(!s.endsWith(".ser")){
+            JOptionPane.showMessageDialog(mui, "File name should end with .ser");
+            return;
+        }
+        File[] fileList = (new File(".")).listFiles((File pathname) -> pathname.getName().endsWith(".ser"));
+        boolean flag = false;
+        for(File f : fileList){
+            if(f.getName().matches(s)){
+                flag = true;
+            }
+        }
+        if(flag){
+            int q = JOptionPane.showConfirmDialog(mui, s + " already exists:\nAre you sure you want to overwrite?");
+            if(q!=0)
+                return;
+        }
+        try {
+            SerializationUtil.serialize(a, s);
+        } catch (IOException e) {
+            return;
+        }
+        JOptionPane.showMessageDialog(mui, s + " saved successfully");
+    }
+
 }
