--- src/contactmanagementsoftware/command/ExitReceiver.java
@@ -1,8 +1,0 @@
-package contactmanagementsoftware.command;
-
-public class ExitReceiver {
-
-    public void exit(){
-        System.exit(0);
-    }
-}
