--- src/contactmanagementsoftware/command/SearchReceiver.java
@@ -1,28 +1,0 @@
-package contactmanagementsoftware.command;
-
-import javax.swing.*;
-
-public class SearchReceiver extends Receiver {
-
-    public void search() {
-        JTextPane details = mui.getDetails();
-
-        String s = (String) JOptionPane.showInputDialog(
-                mui,
-                "Enter name: ",
-                "Input",
-                JOptionPane.PLAIN_MESSAGE,
-                null,
-                null,
-                "");
-        if(s==null)
-            return;
-        jPanel1.setVisible(false);
-        jPanel2.setVisible(true);
-        mui.setStr(s);
-        details.setContentType( "text/html" );
-        mui.runn();
-
-    }
-
-}
